Move featureInfo and its table components inside details-1, more cleaning
