+++ packages/geoview-core/src/core/components/details-1/feature-info-table.tsx
@@ -1,8 +1,7 @@
-/* eslint-disable @typescript-eslint/no-unused-vars */
 import React, { useState } from 'react';
 import linkifyHtml from 'linkify-html';
 import { useTranslation } from 'react-i18next';
-import { Table, TableBody, TableCell, TableContainer, TableRow, Paper } from '@mui/material';
+import { Table, TableBody, TableCell, TableContainer, TableRow } from '@mui/material';
 import { TypeFieldEntry } from '@/api/events/payloads';
 import { LightboxImg, LightBoxSlides } from '../lightbox/lightbox';
 import { CardMedia, Box } from '@/ui';
@@ -14,28 +13,31 @@ interface FeatureInfoTableProps {
   featureInfoList: TypeFieldEntry[];
 }
 
-const sxClasses = {
-  featureInfoItemImage: {
-    cursor: 'pointer',
-  },
-  featureInfoItemValue: {
-    fontSize: '0.85em',
-    marginRight: 0,
-    marginTop: '5px',
-    wordBreak: 'break-word',
-    overflow: 'hidden',
-    textOverflow: 'ellipsis',
-  },
+const getSxClasses = (isOdd = true, isFeatureImage = false) => {
+  return {
+    featureInfoItemValue: {
+      fontSize: '0.85em',
+      marginRight: 0,
+      marginTop: '5px',
+      wordBreak: 'break-word',
+      overflow: 'hidden',
+      textOverflow: 'ellipsis',
+      ...(isFeatureImage && { cursor: 'pointer' }),
+    },
+    featureTableRow: {
+      ...(isOdd && { backgroundColor: '#F1F2F5' }),
+    },
+  };
 };
 
 /**
  * Feature info table that creates a table keys/values of the given feature info
  *
+ * @param {FeatureInfoTableProps} Feature info table properties
  * @returns {JSX.Element} the layers list
  */
 export function FeatureInfoTable({ featureInfoList }: FeatureInfoTableProps): JSX.Element {
   const { t } = useTranslation<string>();
-  // lightbox component state
   const [isLightBoxOpen, setIsLightBoxOpen] = useState(false);
   const [slides, setSlides] = useState<LightBoxSlides[]>([]);
   const [slidesIndex, setSlidesIndex] = useState(0);
@@ -67,7 +69,7 @@ export function FeatureInfoTable({ featureInfoList }: FeatureInfoTableProps): JS
         element = (
           <CardMedia
             key={generateId()}
-            sx={[sxClasses.featureInfoItemValue, sxClasses.featureInfoItemImage]}
+            sx={getSxClasses(true, true).featureInfoItemValue}
             alt={alias}
             src={item}
             tabIndex={0}
@@ -86,7 +88,7 @@ export function FeatureInfoTable({ featureInfoList }: FeatureInfoTableProps): JS
         );
       } else {
         element = (
-          <Box key={generateId()} sx={sxClasses.featureInfoItemValue}>
+          <Box key={generateId()} sx={getSxClasses().featureInfoItemValue}>
             <HtmlToReact htmlContent={sanitizeHtmlContent(linkifyHtml(item, linkifyOptions))} />
           </Box>
         );
@@ -95,9 +97,7 @@ export function FeatureInfoTable({ featureInfoList }: FeatureInfoTableProps): JS
       return element;
     }
 
-    // item must be a string
-    const { alias } = featureInfoItem;
-    const { value } = featureInfoItem;
+    const { alias, value } = featureInfoItem;
     let values: string | string[] = Array.isArray(value) ? String(value.map(stringify)) : String(stringify(value));
     values = values.toString().split(';');
     const results = Array.isArray(values)
@@ -126,22 +126,17 @@ export function FeatureInfoTable({ featureInfoList }: FeatureInfoTableProps): JS
       <Table
         sx={{
           border: '1px solid #BDBDBD',
-          '& th, & td': {
-            // borderBottom: '1px solid red',
-          },
           minWidth: 300,
         }}
       >
         <TableBody>
           {featureInfoList.map((featureInfoItem, index) => (
             // eslint-disable-next-line react/no-array-index-key
-            <TableRow key={index}>
+            <TableRow key={index} sx={getSxClasses(index % 2 > 0).featureTableRow}>
               <TableCell
                 sx={{
                   borderRight: '1px solid #BDBDBD',
                   p: '5px',
-                  // paddingLeft: '8px',
-                  // paddingRight: '50px',
                   width: '70%',
                 }}
               >
@@ -150,7 +145,6 @@ export function FeatureInfoTable({ featureInfoList }: FeatureInfoTableProps): JS
               <TableCell
                 sx={{
                   p: '5px',
-                  // padding: '8px',
                   textAlign: 'start',
                 }}
               >
